*** empty log message ***
git-svn-id: https://nlg0.isi.edu/svn/sbmt/trunk/graehl@649 7f356e55-ce0a-0410-be72-bc6dfe96aa1c

+++ shared/config.h
@@ -7,7 +7,7 @@
 
 #define TREE_SINGLETON_OPT
 typedef short unsigned var_type; // 0 = no var, 1,2,3,... = var index
-typedef int rank_type; // (rank=#children) -1 = any rank, -2 = any tree ... (can't be unsigned type)
+typedef short rank_type; // (rank=#children) -1 = any rank, -2 = any tree ... (can't be unsigned type)
 
 #define COPYRIGHT_YEAR 2004
